Add auto .coveragerc creation and discovery
When target 'coverage' ran with source files in the root directory, it
included all the virtual environment packages as well.  This is not
expected behavior.  Typically, your would run it, discover your mistake,
look at the internal notes and fix it.  Not good.

Now it will auto create a .coveragerc file with directions to omit the
ENV directory and stop the make process. When it runs the second time
the coverage arguments COVER_ARG will contain the config file.

Test files and documentation are updated.