Add 'venv' as virtual environment directory
I've gone through a change in naming virtual environments.  I started
with a hidden directory '.env' and then moved to 'env'.  Now I am
changing to 'venv' for two reasons.  One, it better expresses purpose
not to be confused with environment variables.  Second, it comes later
in the alphabet making it show up later in a directory listing.